+++ spec/controllers/api/grade_entry_forms_controller_spec.rb
@@ -314,7 +314,7 @@ describe Api::GradeEntryFormsController do
       end
     end
     context 'DELETE Destroy' do
-      it 'does not delete a non-existing assignment' do
+      it 'does not delete a non-existing grade entry form' do
         delete :destroy, params: { course_id: course.id, id: -1 }
         expect(response).to have_http_status(404)
       end